pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Proyecto de aplicación web diseñado para que los invitados de un evento, como un cumpleaños o una fiesta, puedan compartir sus fotos fácilmente y verlas en una presentación en tiempo real."
readme = "README.md"
requires-python = ">=3.11.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "certifi==2025.10.5",
    "click==8.3.0",
    "colorama==0.4.6",
    "fastapi==0.119.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "iniconfig==2.1.0",
    "packaging==25.0",
    "piexif==1.1.3",
    "pillow==12.0.0",
    "pluggy==1.6.0",
    "pydantic==2.12.2",
    "pydantic-core==2.41.4",
    "pygments==2.19.2",
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.3",
    "sniffio==1.3.1",
    "starlette==0.48.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn[standard]==0.37.0",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
]
